Just realized that my shaders were in my project, but not in the local folder or in git. Sorry about that.
This is a mismatch between Cherno and my Udemy class. Finally have that
multicolored triangle. Whooooooo... It was YEARS ago when I tried to get
this far. Sick. Messy. Sad. But Colorful and not a bad place to start
with a working minimal triangle.

Funny how Cherno goes straight to uniforms which seem like the "nice"
way to do this, but a toehold is all I usually need.