xcode does not play nicely with git - at least not intuitively. The folder/directory interaction is enough to drive someone insane too. Place something in the right directory and not have it show up in the folder. Not nice. Change the file and not have it show up in git status. Doubly not nice. This is probably caused by misunderstanding of some geewhiz feature, but did it really have to be this way?
It appears that the wrong stb_image.h might have been checked in before
so the previous commit might not be able to be compiled. Sorry. This is
scratch for me and just a mishmash of Cherno's and Hussain's work since
I'm just learning, not trying to perfect anything.

The picture is of Tianjin, one of my homes away from home.